requirements.txt
streamlit
wordcloud
matplotlib
pandas
openpyxl
yfinance
plotly
mplfinance
azure-functions
